conformance GOV: translate and adjust
