+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway.hs
@@ -3,4 +3,5 @@ module Test.Cardano.Ledger.Conformance.SpecTranslate.Conway () where
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Base ()
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Cert ()
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Deleg ()
+import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Gov ()
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Pool ()
